play samsungani if dalvik-cache might have been cleared to avoid long, scary black screen after custom bootanimation stopped and dalvik-cache is regenerated.

+++ sbin/bootanimation.sh
@@ -1,12 +1,18 @@
 #!/sbin/busybox_disabled sh
-# play android bootanimation.zip if it exists, if not play samsungani
+# if dalvik-cache 'seems' to be cleared play samsung animation else
+# play if exist: bootanimation.zip, bootanimation.bin else
+# play samsung animation...
 
-if [ -f /data/local/bootanimation.bin ]; then
-  /data/local/bootanimation.bin
-elif [ -f /data/local/bootanimation.zip ] || [ -f /system/media/bootanimation.zip ]; then
-	exec /sbin/bootanimation &
-    sleep 12
-    kill $!
+if test -e /data/dalvik-cache/system@app@SetupWizard.apk@classes.dex; then 
+    if [ -f /data/local/bootanimation.bin ]; then
+      /data/local/bootanimation.bin
+    elif [ -f /data/local/bootanimation.zip ] || [ -f /system/media/bootanimation.zip ]; then
+        exec /sbin/bootanimation &
+        sleep 12
+        kill $!
+    else
+      /system/bin/samsungani
+    fi
 else
-  /system/bin/samsungani
-fi;
+    /system/bin/samsungani
+fi
